refactor(stats): make residual_error the single owner of the variance math
Relocates the `ErrorSpec` residual-variance-and-derivatives math out of the
~6.5k-line `types.rs` god-file and into `stats/residual_error.rs`, which
already holds the primitive it calls (`residual_variance`) and every consumer
(`compute_r_diag*`, `compute_r_matrix*`, `compute_dr_df_matrices`, IWRES/CWRES).

The math already had a single *logical* owner (`ErrorSpec`) but the wrong
*physical* home, which forced a bidirectional module dependency:
`residual_error.rs` imported `ErrorSpec` from `types.rs`, while
`types.rs::variance_at` reached back for `crate::stats::residual_error::
residual_variance`. That cycle is now dissolved — the boundary is one-way and
`types.rs` no longer references `stats::residual_error` at all.

Two changes, both pure code motion:

* Move the `impl ErrorSpec` numeric block (`sigma_loadings`/`_slopes`,
  `variance_at{,_scaled,_with_correlations}`, `dvar_df{,_scaled}`,
  `d2var_df2{,_scaled}`, `dvar_dlogsigma`, `prop_sigma_slot`, `sigma_types`,
  `combined_additive_sigma_indices`, `has_f_dependent_variance`) verbatim into
  `residual_error.rs`. Relocating an inherent impl is compile-time-only: Rust
  resolves inherent impls from any module of the defining crate, so all call
  sites produce byte-identical code. `variance_at` now calls the same-module
  `residual_variance` directly.
* Promote `variance_rd`/`variance_rd2` from `sens_outer_gradient.rs` into
  `residual_error.rs` as `pub(crate) residual_rd`/`residual_rd2` — the
  canonical scalar (r,d[,d2]) accessors, next to the math they dispatch to.
  All 9 call sites are a mechanical rename with identical arguments.

`types.rs` keeps the `ErrorSpec` data definition and `obs_key`/`obs_keys`
(data/dispatch concerns), plus `CompiledModel::residual_variance_at{,_scaled}`
as thin facades binding `self.residual_correlations`.

Both association traps documented in the code are preserved: `variance_at`
(legacy `(f*s)*(f*s)`) is NOT collapsed into `variance_at_scaled`
(`((f*f)*s)*s`) — they differ by ~1 ULP on ~55% of prop/combined rows and every
bare-sigma R/OFV/CWRES is pinned to the legacy association; and `ruv_scale` is
NOT folded into `residual_rd`, since likelihood.rs scales the variance directly
while sens_outer_gradient scales an FD difference.

Verified bit-identical by diffing the relocated blocks against the originals:
470/470 lines verbatim. `cargo test --lib`: 2570 passed, 0 failed, 0 warnings.

Behavior change: none.
