+++ src/estimation/sens_outer_gradient.rs
@@ -39,6 +39,7 @@ use crate::estimation::parameterization::{
     unpack_params,
 };
 use crate::sens::provider::{subject_sensitivities, ObsSens, SubjectSens};
+use crate::stats::residual_error::{residual_rd, residual_rd2};
 use crate::stats::special::m3_censored_outer;
 use crate::types::{CompiledModel, ModelParameters, Population, Subject};
 use nalgebra::{DMatrix, DVector};
@@ -676,53 +677,6 @@ fn mag_alpha_dtheta(et: &ErrTerms, m: usize) -> f64 {
         + ((r - eps * eps) * inv_r2) * d_th
 }
 
-/// Residual variance `R` and its `f`-derivative `d = ∂R/∂f` for one observation,
-/// dispatching on whether a custom σ-magnitude is active (`mult_row = Some`): the
-/// scaled closed forms when it is, the legacy ones when it is not (`None` keeps
-/// the exact `variance_at`/`dvar_df` association bit-for-bit — the `_scaled`
-/// variants reassociate the `f`-dependent term by ~1 ULP). Callers apply
-/// `ruv_scale` (and any FD combination) at the call site so float association is
-/// unchanged. Diagonal-`R` only (`block_sigma` correlations force FD upstream).
-#[inline]
-fn variance_rd(
-    es: &crate::types::ErrorSpec,
-    cmt: usize,
-    f: f64,
-    sigma: &[f64],
-    mult_row: Option<&[f64]>,
-) -> (f64, f64) {
-    match mult_row {
-        Some(m) => (
-            es.variance_at_scaled(cmt, f, sigma, &[], m),
-            es.dvar_df_scaled(cmt, f, sigma, m),
-        ),
-        None => (es.variance_at(cmt, f, sigma), es.dvar_df(cmt, f, sigma)),
-    }
-}
-
-/// [`variance_rd`] plus the second `f`-derivative `d2 = ∂²R/∂f²`.
-#[inline]
-fn variance_rd2(
-    es: &crate::types::ErrorSpec,
-    cmt: usize,
-    f: f64,
-    sigma: &[f64],
-    mult_row: Option<&[f64]>,
-) -> (f64, f64, f64) {
-    match mult_row {
-        Some(m) => (
-            es.variance_at_scaled(cmt, f, sigma, &[], m),
-            es.dvar_df_scaled(cmt, f, sigma, m),
-            es.d2var_df2_scaled(cmt, sigma, m),
-        ),
-        None => (
-            es.variance_at(cmt, f, sigma),
-            es.dvar_df(cmt, f, sigma),
-            es.d2var_df2(cmt, sigma),
-        ),
-    }
-}
-
 /// The σ-EBE-response data-term coefficient derivative for one observation:
 /// `∂α/∂σ = [2ε/R² + d(2ε²−R)/R³]·Rσ + [(R−ε²)/R²]·dσ`, given `(r, d, eps)` and the
 /// σ-FD slopes `r_sig = ∂R/∂σ`, `d_sig = ∂d/∂σ`. The σ mirror of [`mag_alpha_dtheta`],
@@ -1000,7 +954,7 @@ pub(crate) fn score_core(
             (Some(v), _) => (v, 0.0, 0.0),
             (None, Some((rv, dv, d2v))) => (rv[j], dv[j], d2v[j]),
             (None, None) => {
-                let (r, d, d2) = variance_rd2(&model.error_spec, cmt, f, sigma, mult_row);
+                let (r, d, d2) = residual_rd2(&model.error_spec, cmt, f, sigma, mult_row);
                 (r * ruv_scale, d * ruv_scale, d2 * ruv_scale)
             }
         };
@@ -1718,8 +1672,8 @@ fn sigma_block(
                 let (vp, vm, dp_var, dm_var) = match (&corr_sp, &corr_sm) {
                     (Some((rvp, dvp)), Some((rvm, dvm))) => (rvp[j], rvm[j], dvp[j], dvm[j]),
                     _ => {
-                        let (vp, dp_var) = variance_rd(&model.error_spec, cmt, f, &sp, mult_row);
-                        let (vm, dm_var) = variance_rd(&model.error_spec, cmt, f, &sm, mult_row);
+                        let (vp, dp_var) = residual_rd(&model.error_spec, cmt, f, &sp, mult_row);
+                        let (vm, dm_var) = residual_rd(&model.error_spec, cmt, f, &sm, mult_row);
                         (vp, vm, dp_var, dm_var)
                     }
                 };
@@ -2232,7 +2186,7 @@ pub fn subject_packed_gradient_foce(
         let (r, dd) = match (frem_r0, &corr_rd0) {
             (Some(v), _) => (v, 0.0),
             (None, Some((rv, dv, _))) => (rv[j], dv[j]),
-            (None, None) => variance_rd(&model.error_spec, cmt, f0act, sigma, mult_row),
+            (None, None) => residual_rd(&model.error_spec, cmt, f0act, sigma, mult_row),
         };
         if !(r.is_finite() && r > 0.0) {
             return None;
@@ -2616,8 +2570,8 @@ pub fn subject_eta_dx_iov(
             // Magnitude-scaled `∂R/∂σ`,`∂d/∂σ` (consistent with the scaled `et.r`/`et.d`).
             let mult_row: Option<&[f64]> =
                 mult.as_ref().and_then(|mm| mm.get(j)).map(|v| v.as_slice());
-            let (var_p, dvar_p) = variance_rd(&model.error_spec, cmt, f, &sp, mult_row);
-            let (var_m, dvar_m) = variance_rd(&model.error_spec, cmt, f, &sm, mult_row);
+            let (var_p, dvar_p) = residual_rd(&model.error_spec, cmt, f, &sp, mult_row);
+            let (var_m, dvar_m) = residual_rd(&model.error_spec, cmt, f, &sm, mult_row);
             let r_sig = (var_p - var_m) / (2.0 * h);
             let d_sig = (dvar_p - dvar_m) / (2.0 * h);
             let dalpha = dalpha_dsigma(r, d, eps, r_sig, d_sig);
@@ -2738,7 +2692,7 @@ pub fn subject_packed_gradient_foce_iov(
         let cmt = err_keys[j];
         let f0act = sens0.obs[j].f;
         let mult_row: Option<&[f64]> = mult.as_ref().and_then(|m| m.get(j)).map(|v| v.as_slice());
-        let (r, d) = variance_rd(&model.error_spec, cmt, f0act, sigma, mult_row);
+        let (r, d) = residual_rd(&model.error_spec, cmt, f0act, sigma, mult_row);
         if !(r.is_finite() && r > 0.0) {
             return None;
         }
@@ -3087,8 +3041,8 @@ pub fn subject_eta_dx(
                 let (var_p, var_m, dvar_p, dvar_m) = match (&corr_sp, &corr_sm) {
                     (Some((rvp, dvp)), Some((rvm, dvm))) => (rvp[j], rvm[j], dvp[j], dvm[j]),
                     _ => {
-                        let (var_p, dvar_p) = variance_rd(&model.error_spec, cmt, f, &sp, mult_row);
-                        let (var_m, dvar_m) = variance_rd(&model.error_spec, cmt, f, &sm, mult_row);
+                        let (var_p, dvar_p) = residual_rd(&model.error_spec, cmt, f, &sp, mult_row);
+                        let (var_m, dvar_m) = residual_rd(&model.error_spec, cmt, f, &sm, mult_row);
                         (var_p, var_m, dvar_p, dvar_m)
                     }
                 };
